등록 컨트롤러 따로
